Renamed packages, dependencies work to reduce OWASP warnings (#14)
* fixes spotbugs reported errors

* renamed packages, fixes #11

* dependency upgrades, fixes most owasp warning

* fixes jquery dependency warning
